# api/api-docs.yaml
openapi: 3.0.0
info:
  title: Greenalytic User Management API
  version: 1.0.0
  description: API for managing users with authentication and pagination
servers:
  - url: http://localhost:5000/greenalytic
    description: Local Development Server
  - url: https://greenalytic-vehicle-monitoring-1.onrender.com/greenalytic
    description: Production Server

# Define Authentication Scheme
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT  # JSON Web Token (JWT) format
  schemas:
...
    EmissionData:
      type: object
      properties:
        id:
          type: integer
          example: 1
        co2Percentage:
          type: number
          example: 4.5
        coPercentage:
          type: number
          example: 0.3
        o2Percentage:
          type: number
          example: 21.0
        hcPPM:
          type: integer
          example: 200
        vehicleId:
          type: integer
          example: 101
        plateNumber:
          type: string
          example: RAB123X
        trackingDeviceId:
          type: integer
          example: 5001
        timestamp:
          type: string
          format: date-time
          example: "2025-03-01T10:30:00Z"
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        totalCount:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 10
paths:
...
  /emissions:
    post:
      summary: Create a new emission data record
      security:
        - bearerAuth: []
      tags:
        - Emission Data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmissionData"
      responses:
        "201":
          description: Emission data created successfully
        "400":
          description: Bad Request - Missing required fields
        "500":
          description: Internal Server Error

    get:
      summary: Get all emission data (paginated)
      security:
        - bearerAuth: []
      tags:
        - Emission Data
      parameters:
        - name: page
          in: query
          schema:
            type: integer
          example: 1
        - name: limit
          in: query
          schema:
            type: integer
          example: 10
        - name: startTime
          in: query
          schema:
            type: string
            format: date-time
        - name: endTime
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: List of emission data
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/EmissionData"
                  meta:
                    $ref: "#/components/schemas/PaginationMeta"
        "500":
          description: Internal Server Error
  /emissions/{id}:
    get:
      summary: Get emission data by ID
      security:
        - bearerAuth: []
      tags:
        - Emission Data
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Emission data found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmissionData"
        "404":
          description: Emission data not found
        "500":
          description: Internal Server Error

    put:
      summary: Update emission data by ID
      security:
        - bearerAuth: []
      tags:
        - Emission Data
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmissionData"
      responses:
        "200":
          description: Emission data updated successfully
        "404":
          description: Emission data not found
        "500":
          description: Internal Server Error

    delete:
      summary: Delete emission data by ID
      security:
        - bearerAuth: []
      tags:
        - Emission Data
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Emission data deleted successfully
        "404":
          description: Emission data not found
        "500":
          description: Internal Server Error
  /emissions/vehicle/{vehicleId}:
    get:
      summary: Get emission data for a specific vehicle
      description: Retrieves emission data for a given vehicle with pagination, filtering by date range if provided.
      tags:
        - Emission Data
      parameters:
        - in: path
          name: vehicleId
          required: true
          schema:
            type: integer
          description: The ID of the vehicle.
        - in: query
          name: page
          required: false
          schema:
            type: integer
            default: 1
          description: The page number for pagination (default is 1).
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            default: 10
          description: The number of items per page (default is 10).
        - in: query
          name: startTime
          required: false
          schema:
            type: string
            format: date-time
          description: The start timestamp for filtering emission data (optional).
        - in: query
          name: endTime
          required: false
          schema:
            type: string
            format: date-time
          description: The end timestamp for filtering emission data (optional).
      responses:
        '200':
          description: A list of emission data with pagination metadata.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          description: The ID of the emission record.
                        vehicleId:
                          type: integer
                          description: The ID of the vehicle.
                        timestamp:
                          type: string
                          format: date-time
                          description: The timestamp when the emission data was recorded.
                        trackingDevice:
                          type: object
                          properties:
                            id:
                              type: integer
                            name:
                              type: string
                          description: The tracking device associated with the emission data.
                  meta:
                    type: object
                    properties:
                      currentPage:
                        type: integer
                        description: The current page number.
                      totalPages:
                        type: integer
                        description: The total number of pages.
                      remainingItems:
                        type: integer
                        description: The number of items remaining after the current page.
                      totalItems:
                        type: integer
                        description: The total number of items matching the query.
                      limit:
                        type: integer
                        description: The number of items per page.
        '400':
          description: Invalid parameters.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: The error message explaining what was wrong with the request.
        '500':
          description: Failed to fetch emission data.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: An error message explaining the failure.
    /vehicles/addvehicletouser/{userId}:
    post:
      summary: Add a Vehicle to a User
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                plateNumber:
                  type: string
                  example: "ABC123"
                chassisNumber:
                  type: string
                  example: "CHS12345678"
                vehicleType:
                  type: string
                  example: "SUV"
                vehicleModel:
                  type: string
                  example: "Toyota Prado"
                yearOfManufacture:
                  type: integer
                  example: 2020
                usage:
                  type: string
                  example: "Personal"
      responses:
        "201":
          description: Vehicle added successfully
        "400":
          description: Bad request
        "500":
          description: Internal server error
...